+++ bidict/__init__.py
@@ -18,7 +18,7 @@ from ._bidict import bidict
 from ._loose import loosebidict
 from ._frozen import frozenbidict
 from ._named import namedbidict
-from ._ordered import (OrderedBidirectionalMapping,
+from ._ordered import (OrderedMapping, OrderedBidirectionalMapping,
                        orderedbidict, frozenorderedbidict, looseorderedbidict)
 from .util import pairs, inverted
 
@@ -37,6 +37,7 @@ __all__ = (
     'loosebidict',
     'frozenbidict',
     'namedbidict',
+    'OrderedMapping',
     'OrderedBidirectionalMapping',
     'orderedbidict',
     'frozenorderedbidict',